+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+<a name="0.0.2"></a>
+# 0.0.2 (2019-06-30)
+
+### Bug Fixes
+- fixed error in `MixpanelHttpTracker` middleware
+
+### Feature Added
+- added `trackUserDeletion` method to instance
+
 <a name="0.0.1"></a>
 # 0.0.1 (2019-06-28)